Remove useless tests. Move Maps activity to its own package.
