+++ app/src/androidTest/java/com/github/leuludyha/ibdb/GoogleMapsActivityTest.kt
@@ -1,6 +1,7 @@
 package com.github.leuludyha.ibdb
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import com.github.leuludyha.ibdb.maps.GoogleMapsActivity
 
 import org.junit.Rule
 import org.junit.Test
@@ -11,7 +12,16 @@ class GoogleMapsActivityTest {
     val composeTestRule = createAndroidComposeRule<GoogleMapsActivity>()
 
     @Test
-    fun clickingOnEPFLMarkerShowsPreviouslyInvisibleDelimiterPolygon() {
-        //TODO: I don't see how to do this actually...
+    fun infoWindowWithEPFLTagDoesNotExistByDefault() {
+        //TODO: Useless test, remove
+        //composeTestRule.onNodeWithTag("info_window_epfl").assertDoesNotExist()
+    }
+
+    @Test
+    fun infoWindowWithEPFLTagAppearsAfterClickOnMarker() {
+        //TODO: I couldn't find any way to test the markers
+//        composeTestRule.onNodeWithTag("test_tag").performClick()
+//        composeTestRule.onNodeWithTag("GoogleMap").onChild().performClick();
+//        composeTestRule.onNodeWithTag("info_window_epfl").assertExists()
     }
 }
